+++ scilab/modules/sparse/sci_gateway/cpp/sparse_gw.cpp
@@ -39,5 +39,6 @@ int SparseModule::Load()
     symbol::Context::getInstance()->addFunction(types::Function::createFunction(L"bfinit", &sci_bfinit, MODULE_NAME));
     symbol::Context::getInstance()->addFunction(types::Function::createFunction(L"inpnvi", &sci_inpnv, MODULE_NAME));
     symbol::Context::getInstance()->addFunction(types::Function::createFunction(L"blkfc1i", &sci_blkfc1i, MODULE_NAME));
+    symbol::Context::getInstance()->addFunction(types::Function::createFunction(L"blkslvi", &sci_blkslvi, MODULE_NAME));
     return 1;
 }